[FE] rename: nav 컴포넌트 이동
